Table.copy() method works now. Solved some details that allow to compile now under MacOSX. All test passes under MacOSX.
git-svn-id: http://www.pytables.org/svn/pytables/trunk@240 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94

+++ src/getfieldfmt.c
@@ -108,7 +108,7 @@ int format_element(hid_t type_id,
     }
     break; /* case H5T_FLOAT */
   case H5T_STRING:                  /* char or string */
-    snprintf(temp, 255, "a%d,", member_size);  /* Always a CharArray */
+    snprintf(temp, 255, "a%d,", (int)member_size);  /* Always a CharArray */
     PyList_Append(types, PyString_FromString("a"));
     strcat( format, temp );       /* string */
     break; /* case H5T_STRING */
